chore(refactor): moves generic node components to global location
