+++ frontend/packages/syntara-ui/src/components/nodes/NodeBody.test.tsx
