+++ frontend/packages/syntara-ui/src/components/nodes/NodeExpandedContext.tsx
