+++ frontend/packages/syntara-ui/src/components/nodes/NodeMenu.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { describe, expect, it, vi } from 'vitest'
 
-import type { NodeMenuAction } from '../hooks/useNodeMenuActions'
+import type { NodeMenuAction } from '../../routes/workflows/canvas/nodes/hooks/useNodeMenuActions'
 
 import { NodeMenu } from './NodeMenu'
