+++ frontend/packages/syntara-ui/src/components/nodes/NodeTitle.tsx
