+++ frontend/packages/syntara-ui/src/routes/builder/BuilderContent.tsx
@@ -10,6 +10,7 @@ import { executionsClient, workflowClient } from '../../client'
 import { SynPage } from '../../components/layout/SynPage'
 import { SynPanel } from '../../components/layout/SynPanel'
 import { SynReactFlowViewportGuard } from '../../components/layout/SynReactFlowViewportGuard'
+import { NodeExpandedAllContext } from '../../components/nodes/NodeExpandedAllContext'
 import { useSearchParams } from '../../hooks/routing/useSearchParams'
 import { useCursorPagination } from '../../hooks/useCursorPagination'
 import { useProjectSelector } from '../../hooks/useProjectSelector'
@@ -19,7 +20,6 @@ import { getErrorMessage } from '../../utils/apiErrors'
 import { detachPromise } from '../../utils/detachPromise'
 import { ApprovalSidePanel } from '../executions/ApprovalSidePanel'
 import { transformExecutionStatusFilter } from '../executions/executionFilters'
-import { NodeExpandedAllContext } from '../workflows/canvas/nodes/common/NodeExpandedAllContext'
 
 import styles from './BuilderContent.module.css'
 import { BuilderFlow } from './BuilderFlow'
