+++ frontend/packages/syntara-ui/src/routes/workflows/canvas/nodes/GenericNode.tsx
@@ -3,14 +3,14 @@ import { RhUiSettingsIcon } from '@patternfly/react-icons'
 import type { TaskActivity } from '@syntara/contracts'
 import { type Node, type NodeProps } from '@xyflow/react'
 
+import { NodeBody } from '../../../../components/nodes/NodeBody'
+import { NodeComponent } from '../../../../components/nodes/NodeComponent'
 import { FlowNodeType } from '../../../../constants'
 import { getActivityMetadata } from '../../../../stores/useWorkflowStore'
 import type { ActivityStatus } from '../../execution/types'
 import { getNodeTypeColor } from '../nodeTypeColors'
 import { semanticZoomActivityTitle } from '../semanticZoom'
 
-import { NodeBody } from './common/NodeBody'
-import { NodeComponent } from './common/NodeComponent'
 import { StandardNodeHeader } from './common/StandardNodeHeader'
 import { renderNodeIcon } from './renderNodeIcon'
