+++ frontend/packages/syntara-ui/src/routes/workflows/canvas/nodes/TriggerNode.tsx
@@ -3,6 +3,10 @@ import { TriggerTypeEnum } from '@syntara/contracts'
 import { type Node, type NodeProps, useStore } from '@xyflow/react'
 import type { CSSProperties } from 'react'
 
+import { NodeBody } from '../../../../components/nodes/NodeBody'
+import { NodeComponent } from '../../../../components/nodes/NodeComponent'
+import { NodeHeader } from '../../../../components/nodes/NodeHeader'
+import { NodeMenu } from '../../../../components/nodes/NodeMenu'
 import { FlowNodeType } from '../../../../constants'
 import { parseTriggerIndex } from '../../../../utils/triggerNodeIds'
 import { useIsActiveExecution } from '../../../builder/ActiveExecutionContext'
@@ -12,10 +16,6 @@ import type { ActivityStatus } from '../../execution/types'
 import { getNodeTypeColor } from '../nodeTypeColors'
 import { semanticZoomActivityTitle } from '../semanticZoom'
 
-import { NodeBody } from './common/NodeBody'
-import { NodeComponent } from './common/NodeComponent'
-import { NodeHeader } from './common/NodeHeader'
-import { NodeMenu } from './common/NodeMenu'
 import { MenuNodeType, useNodeMenuActions } from './hooks/useNodeMenuActions'
 import { nodeMetadata } from './nodeMetadata'
 import { renderNodeIcon } from './renderNodeIcon'
