+++ frontend/packages/syntara-ui/src/routes/workflows/canvas/nodes/WaitNode.tsx
@@ -3,6 +3,8 @@ import { type Node, type NodeProps } from '@xyflow/react'
 import { useShallow } from 'zustand/react/shallow'
 
 import { SynDetailList } from '../../../../components/details/SynDetailList'
+import { NodeBody } from '../../../../components/nodes/NodeBody'
+import { NodeComponent } from '../../../../components/nodes/NodeComponent'
 import { RegistryNodeId } from '../../../../constants'
 import { formatDurationLabel } from '../../../builder/utils/timeUtils'
 import type { ActivityStatus } from '../../execution/types'
@@ -11,8 +13,6 @@ import { getNodeTypeColor } from '../nodeTypeColors'
 import { semanticZoomActivityTitle } from '../semanticZoom'
 
 import { renderText } from './common/detailRenderers'
-import { NodeBody } from './common/NodeBody'
-import { NodeComponent } from './common/NodeComponent'
 import { StandardNodeHeader } from './common/StandardNodeHeader'
 import { MenuNodeType, useNodeMenuActions } from './hooks/useNodeMenuActions'
 import { useWaitCountdown } from './hooks/useWaitCountdown'
